Move isRequestForHtml to BootstrapHandler
