+++ flow-server/src/main/java/com/vaadin/flow/server/communication/IndexHtmlRequestHandler.java
@@ -40,7 +40,6 @@ import com.vaadin.flow.internal.springcsrf.SpringCsrfTokenUtil;
 import com.vaadin.flow.server.AppShellRegistry;
 import com.vaadin.flow.server.BootstrapHandler;
 import com.vaadin.flow.server.Constants;
-import com.vaadin.flow.server.HandlerHelper;
 import com.vaadin.flow.server.VaadinContext;
 import com.vaadin.flow.server.VaadinRequest;
 import com.vaadin.flow.server.VaadinResponse;
@@ -233,10 +232,6 @@ public class IndexHtmlRequestHandler extends JavaScriptBootstrapHandler {
                 && !BootstrapHandler.isVaadinStaticFileRequest(request)
                 && request.getService().getBootstrapUrlPredicate()
                         .isValidUrl(request);
-    }
-
-    protected boolean isRequestForHtml(VaadinRequest request) {
-        return HandlerHelper.isRequestForHtml(request);
     }
 
     @Override
